Specify the number of fireworks to start
This is different than launch() because it respects the delay (whereas launch() just does everything at once).

Feel free to not accept this or change it however you like. I haven't tested it at all as the changes I made were in JavaScript and I can't get the project to build/work with typescript.